Reconfigure test containers for IT and improve testing

+++ src/main/java/com/raffleease/raffleease/Domains/Images/Repository/ImagesRepository.java
@@ -14,7 +14,7 @@ import java.util.List;
 @Repository("ImagesRepository")
 public interface ImagesRepository extends JpaRepository<Image, Long> {
     @Query("SELECT COUNT(i) FROM Image i WHERE i.raffle IS NULL AND i.association = :association")
-    long countPendingImagesByAssociation(@Param("association") Association association);
+    int countPendingImagesByAssociation(@Param("association") Association association);
 
     List<Image> findAllByRaffleIsNullAndCreatedAtBefore(LocalDateTime cutoff);
     List<Image> findAllByRaffleIsNullAndAssociation(Association association);
